Redesign Collector's ZookeeperCluster logic. #1305
1. [collector] changed zookeeper's job scheduler.
2. [web] apply spring bean in cluster package.
