+++ collector/src/main/java/com/navercorp/pinpoint/collector/cluster/zookeeper/ZookeeperProfilerClusterManager.java
@@ -16,30 +16,25 @@
 
 package com.navercorp.pinpoint.collector.cluster.zookeeper;
 
-import java.nio.charset.Charset;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-
-import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import com.navercorp.pinpoint.collector.cluster.ClusterPointRepository;
 import com.navercorp.pinpoint.collector.cluster.PinpointServerClusterPoint;
 import com.navercorp.pinpoint.collector.cluster.WorkerState;
 import com.navercorp.pinpoint.collector.cluster.WorkerStateContext;
-import com.navercorp.pinpoint.collector.cluster.zookeeper.job.DeleteJob;
-import com.navercorp.pinpoint.collector.cluster.zookeeper.job.UpdateJob;
 import com.navercorp.pinpoint.collector.receiver.tcp.AgentHandshakePropertyType;
 import com.navercorp.pinpoint.rpc.common.SocketStateCode;
 import com.navercorp.pinpoint.rpc.server.PinpointServer;
 import com.navercorp.pinpoint.rpc.server.handler.ServerStateChangeEventHandler;
 import com.navercorp.pinpoint.rpc.util.MapUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.nio.charset.Charset;
+import java.util.List;
+import java.util.Map;
 
 /**
- * @author koo.taejin
+ * @Author Taejin Koo
  */
 public class ZookeeperProfilerClusterManager implements ServerStateChangeEventHandler {
 
@@ -49,11 +44,13 @@ public class ZookeeperProfilerClusterManager implements ServerStateChangeEventHa
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-    private final ZookeeperLatestJobWorker worker;
+    private final ZookeeperJobWorker worker;
 
     private final WorkerStateContext workerState;
 
     private final ClusterPointRepository profileCluster;
+
+    private final Object lock = new Object();
 
     // keep it simple - register on RUN, remove on FINISHED, skip otherwise
     // should only be instantiated when cluster is enabled.
@@ -61,29 +58,27 @@ public class ZookeeperProfilerClusterManager implements ServerStateChangeEventHa
         this.workerState = new WorkerStateContext();
         this.profileCluster = profileCluster;
 
-        this.worker = new ZookeeperLatestJobWorker(client, serverIdentifier);
+        this.worker = new ZookeeperJobWorker(client, serverIdentifier);
     }
 
     public void start() {
         switch (this.workerState.getCurrentState()) {
             case NEW:
                 if (this.workerState.changeStateInitializing()) {
-                    logger.info("{} initialization started.", this.getClass().getSimpleName());
+                    logger.info("start() started.");
 
-                    if (worker != null) {
-                        worker.start();
-                    }
-
+                    worker.start();
                     workerState.changeStateStarted();
-                    logger.info("{} initialization completed.", this.getClass().getSimpleName());
+
+                    logger.info("start() completed.");
 
                     break;
                 }
             case INITIALIZING:
-                logger.info("{} already initializing.", this.getClass().getSimpleName());
+                logger.info("start() failed. caused:already initializing.");
                 break;
             case STARTED:
-                logger.info("{} already started.", this.getClass().getSimpleName());
+                logger.info("start() failed. caused:already started.");
                 break;
             case DESTROYING:
                 throw new IllegalStateException("Already destroying.");
@@ -98,24 +93,22 @@ public class ZookeeperProfilerClusterManager implements ServerStateChangeEventHa
         if (!(this.workerState.changeStateDestroying())) {
             WorkerState state = this.workerState.getCurrentState();
 
-            logger.info("{} already {}.", this.getClass().getSimpleName(), state.toString());
+            logger.info("stop() failed. caused:unexpected state.");
             return;
         }
 
-        logger.info("{} destroying started.", this.getClass().getSimpleName());
+        logger.info("stop() started.");
 
-        if (worker != null) {
-            worker.stop();
-        }
-
+        worker.stop();
         this.workerState.changeStateStopped();
-        logger.info("{} destroying completed.", this.getClass().getSimpleName());
+
+        logger.info("stop() completed.");
     }
 
     @Override
     public void eventPerformed(PinpointServer pinpointServer, SocketStateCode stateCode) {
         if (workerState.isStarted()) {
-            logger.info("eventPerformed PinpointServer={}, State={}", pinpointServer, stateCode);
+            logger.info("eventPerformed() started. (PinpointServer={}, State={})", pinpointServer, stateCode);
 
             Map agentProperties = pinpointServer.getChannelProperties();
 
@@ -124,48 +117,41 @@ public class ZookeeperProfilerClusterManager implements ServerStateChangeEventHa
                 return;
             }
 
-            if (SocketStateCode.RUN_DUPLEX == stateCode) {
-                UpdateJob job = new UpdateJob(pinpointServer, new byte[0]);
-                profileCluster.addClusterPoint(new PinpointServerClusterPoint(pinpointServer));
-                worker.putJob(job);
-            } else if (SocketStateCode.isClosed(stateCode)) {
-                DeleteJob job = new DeleteJob(pinpointServer);
-                worker.putJob(job);
-                profileCluster.removeClusterPoint(new PinpointServerClusterPoint(pinpointServer));
+            synchronized (lock) {
+                if (SocketStateCode.RUN_DUPLEX == stateCode) {
+                    profileCluster.addClusterPoint(new PinpointServerClusterPoint(pinpointServer));
+                    worker.addPinpointServer(pinpointServer);
+                } else if (SocketStateCode.isClosed(stateCode)) {
+                    profileCluster.removeClusterPoint(new PinpointServerClusterPoint(pinpointServer));
+                    worker.removePinpointServer(pinpointServer);
+                }
             }
         } else {
             WorkerState state = this.workerState.getCurrentState();
-            logger.info("{} invalid state {}.", this.getClass().getSimpleName(), state.toString());
+            logger.info("eventPerformed() failed. caused:unexpected state.");
             return;
         }
     }
     
     @Override
     public void exceptionCaught(PinpointServer pinpointServer, SocketStateCode stateCode, Throwable e) {
-        logger.warn("ZookeeperProfilerClusterManager exceptionCaught() (pinpointServer:{}, PinpointServerStateCode:{}). Error: {}.", pinpointServer, stateCode, e.getMessage(), e);
+        logger.warn("exceptionCaught(). (pinpointServer:{}, PinpointServerStateCode:{}). caused:{}.", pinpointServer, stateCode, e.getMessage(), e);
     }
-    
-    public List<String> getClusterData() {
-        byte[] contents = worker.getClusterData();
-        if (contents == null) {
-            return Collections.emptyList();
-        }
 
-        List<String> result = new ArrayList<>();
+    public void initZookeeperClusterData() {
+        profileCluster.clear();
 
-        String clusterData = new String(contents, charset);
-        String[] allClusterData = clusterData.split(PROFILER_SEPARATOR);
-        for (String eachClusterData : allClusterData) {
-            if (!StringUtils.isBlank(eachClusterData)) {
-                result.add(eachClusterData);
+        synchronized (lock) {
+            List clusterPointList = profileCluster.getClusterPointList();
+            for (Object clusterPoint : clusterPointList) {
+                if (clusterPoint instanceof PinpointServerClusterPoint) {
+                    PinpointServer pinpointServer = ((PinpointServerClusterPoint) clusterPoint).getPinpointServer();
+                    if (SocketStateCode.isRunDuplex(pinpointServer.getCurrentStateCode())) {
+                        worker.addPinpointServer(pinpointServer);
+                    }
+                }
             }
         }
-
-        return result;
-    }
-
-    public List<PinpointServer> getRegisteredPinpointServerList() {
-        return worker.getRegisteredPinpointServerList();
     }
 
     private boolean skipAgent(Map<Object, Object> agentProperties) {
