+++ web/src/main/java/com/navercorp/pinpoint/web/controller/CommandController.java
@@ -23,7 +23,6 @@ import com.navercorp.pinpoint.thrift.io.HeaderTBaseDeserializer;
 import com.navercorp.pinpoint.thrift.io.HeaderTBaseSerializer;
 import com.navercorp.pinpoint.thrift.io.SerializerFactory;
 import com.navercorp.pinpoint.web.cluster.PinpointRouteResponse;
-import com.navercorp.pinpoint.web.cluster.connection.WebClusterConnectionManager;
 import com.navercorp.pinpoint.web.service.AgentService;
 import com.navercorp.pinpoint.web.vo.AgentInfo;
 import org.apache.commons.lang3.StringUtils;
@@ -56,9 +55,6 @@ public class CommandController {
 
     @Autowired
     private DeserializerFactory<HeaderTBaseDeserializer> commandDeserializerFactory;
-
-    @Autowired
-    private WebClusterConnectionManager clusterConnectionManager;
 
     @Autowired
     private AgentService agentService;
