+++ web/src/main/resources/applicationContext-web.xml
@@ -79,10 +79,18 @@
         <constructor-arg value="#{T(com.navercorp.pinpoint.common.Version).VERSION}" />
     </bean>
 
-    <bean id="clusterConnectionManager" class="com.navercorp.pinpoint.web.cluster.connection.WebClusterConnectionManager">
+    <bean id="clusterConnectionManager" class="com.navercorp.pinpoint.web.cluster.connection.ClusterConnectionManager">
         <constructor-arg ref="config" />
     </bean>
-    
+
+    <bean id="clusterDataManager" class="com.navercorp.pinpoint.web.cluster.zookeeper.ZookeeperClusterDataManager">
+        <constructor-arg ref="config" />
+    </bean>
+
+    <bean id="clusterManager" class="com.navercorp.pinpoint.web.cluster.ClusterManager">
+        <constructor-arg ref="config" />
+    </bean>
+
     <bean id="agentEventMessageDeserializer" class="com.navercorp.pinpoint.common.util.AgentEventMessageDeserializer">
         <constructor-arg ref="commandHeaderTBaseDeserializerFactory"/>
     </bean>
